refactor : DB 연결 관련 예외 처리 변경
